backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==11.2.1
# PostgreSQL driver (install only in Linux/macOS or CI). On Windows/Python 3.13 it's problematic.
# psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
supabase==2.4.0
python-magic==0.4.27
drf-yasg==1.21.7
django-storages==1.14.2
requests==2.32.3

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2

